# backend/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Team Manager API
  description: >-
    REST API for the Teamverwaltung sports-club management application.


    Every endpoint may return `429 Too Many Requests` (global per-IP rate
    limit, `RATE_LIMIT_RPS`) or `413 Payload Too Large` (request body cap,
    4 MB by default) — see the `TooManyRequests`/`PayloadTooLarge` reusable
    responses. These aren't repeated on every operation below to keep the
    spec readable; `/auth/login` documents its own stricter rate limit
    explicitly since it differs from the global default.
  version: "1.0.0"
  contact:
    name: Team Manager
    url: https://github.com/yoadey/team-manager
servers:
  - url: /api/v1
    description: Current server
security:
  - bearerAuth: []

# ─── Tags ────────────────────────────────────────────────────────────────────
...
paths:
...
  /teams/{teamId}/members:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [members]
      operationId: listMembers
      summary: List team members
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Member"
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null when there are no more items.
  /teams/{teamId}/members/{membershipId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/membershipId"
    patch:
      tags: [members]
      operationId: updateMember
      summary: Update member profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      tags: [members]
      operationId: removeMember
      summary: Remove member from team
      responses:
        "204":
          description: No Content
  /teams/{teamId}/members/{membershipId}/roles:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/membershipId"
    put:
      tags: [members]
      operationId: setMemberRoles
      summary: Replace member role assignments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRolesRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # ── Roles ───────────────────────────────────────────────────────────────────
...
  /teams/{teamId}/news:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [news]
      operationId: listNews
      summary: List team news
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/NewsItem"
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null when there are no more items.
    post:
      tags: [news]
      operationId: createNews
      summary: Create news item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /teams/{teamId}/news/{newsId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: newsId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [news]
      operationId: updateNews
      summary: Update news item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNewsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      tags: [news]
      operationId: deleteNews
      summary: Delete news item
      responses:
        "204":
          description: No Content

  # ── Polls ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    membershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
...
    cursor:
      name: cursor
      in: query
      description: Opaque keyset-pagination cursor returned as nextCursor by a prior page.
      schema:
        type: string
...
  schemas:
...
    PermLevel:
      type: string
      enum: [none, read, write]
    Permissions:
      type: object
      description: Per-module permission levels
      required: [events, members, finances, news, polls, settings]
      properties:
        events:
          $ref: "#/components/schemas/PermLevel"
        members:
          $ref: "#/components/schemas/PermLevel"
        finances:
          $ref: "#/components/schemas/PermLevel"
        news:
          $ref: "#/components/schemas/PermLevel"
        polls:
          $ref: "#/components/schemas/PermLevel"
        settings:
          $ref: "#/components/schemas/PermLevel"
...
    Role:
      type: object
      required: [id, teamId, name, system, permissions]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        name:
          type: string
        system:
          type: boolean
        color:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
...
    Member:
      type: object
      required: [membershipId, userId, name, email, avatarColor, roles, joinedAt]
      properties:
        membershipId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birthday:
          type: string
          format: date
        address:
          type: string
        avatarColor:
          type: string
        hasPhoto:
          type: boolean
        group:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        primaryRole:
          $ref: "#/components/schemas/Role"
        perms:
          $ref: "#/components/schemas/Permissions"
        joinedAt:
          type: string
          format: date-time
    UpdateMemberRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birthday:
          type: string
          format: date
        address:
          type: string
        roleIds:
          type: array
          items:
            type: string
            format: uuid
        group:
          type: string
    SetRolesRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          items:
            type: string
            format: uuid

    # ── Events ───────────────────────────────────────────────────────────────
...
    NewsItem:
      type: object
      required: [id, teamId, authorId, title, body, pinned, createdAt]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        authorName:
          type: string
        authorColor:
          type: string
        hasAuthorPhoto:
          type: boolean
    CreateNewsRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        body:
          type: string
          minLength: 1
          maxLength: 10000
        pinned:
          type: boolean
          default: false
    UpdateNewsRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        body:
          type: string
          minLength: 1
          maxLength: 10000
        pinned:
          type: boolean

    # ── Polls ────────────────────────────────────────────────────────────────
...